refactor: 패키지 이동 및, Results를 타일로 변경하는 Converter 생성하여 분리
