--- src/main/kotlin/wordle/Results.kt
@@ -1,23 +1,0 @@
-package wordle
-
-import wordle.domain.Result
-
-data class Results(
-    val values: List<Result>
-) {
-    fun convert(): String {
-        val builder = StringBuilder()
-        for (result in values) {
-            builder.append(when (result) {
-                Result.ABSENT -> "⬜"
-                Result.PRESENT -> "🟨"
-                else -> "🟩"
-            })
-        }
-        return builder.toString()
-    }
-
-    fun isAnswer(): Boolean {
-        return values.all { it == Result.CORRECT }
-    }
-}
