+++ src/main/kotlin/wordle/Wordle.kt
@@ -1,5 +1,6 @@
 package wordle
 
+import wordle.domain.Results
 import wordle.domain.Word
 
 class Wordle(
