+++ src/main/kotlin/wordle/WordleResults.kt
@@ -1,5 +1,8 @@
 package wordle
 
+import wordle.domain.Results
+import wordle.translation.ResultsConverter
+
 class WordleResults(
     private val value: MutableList<Results> = mutableListOf()
 ) {
@@ -10,7 +13,7 @@ class WordleResults(
     fun display(): String {
         val stringBuilder: StringBuilder = StringBuilder()
 
-        value.forEach { stringBuilder.append(it.convert()).append("\n") }
+        value.forEach { stringBuilder.append(ResultsConverter.convert(it)).append("\n") }
 
         return stringBuilder.toString().trim()
     }
